Code cleaned
